+++ include/axl/axl-version.h
@@ -28,8 +28,8 @@
 
 #define AXL_VERSION_MAJOR   4
 #define AXL_VERSION_MINOR   3
-#define AXL_VERSION_PATCH   3
-#define AXL_VERSION_STRING  "4.3.3"
+#define AXL_VERSION_PATCH   4
+#define AXL_VERSION_STRING  "4.3.4"
 
 /* Encoded as 0xMMmmpp (major, minor, patch) for simple comparisons. */
 #define AXL_VERSION_NUMBER \